fix: msg id serialization

+++ rosserial_client/src/ros_lib/ros/node_handle.h
@@ -427,8 +427,8 @@ namespace ros {
         message_out[2] = (unsigned char) ((unsigned int)l&255);
         message_out[3] = (unsigned char) ((unsigned int)l>>8);
 	message_out[4] = 255 - ((message_out[2] + message_out[3])%256);
-        message_out[5] = (unsigned char) id&255;
-        message_out[6] = ((unsigned char) id>>8);
+        message_out[5] = (unsigned char) ((int)id&255);
+        message_out[6] = (unsigned char) ((int)id>>8);
 
         /* calculate checksum */
         int chk = 0;
